Refactor wir_optimize: reorganize into focused single-pass modules
Split the ad-hoc wir_optimize module into well-organized files where
each file corresponds to one optimization pass (or a closely related
group). No behavioral changes — golden fixtures are identical.

File changes:
- sroa.rs (2528 lines) → sroa_return.rs + sroa_param.rs
- forward.rs → const_forward.rs + bounds_check.rs
- elide.rs → elide_struct.rs + elide_local.rs
- peephole.rs → peephole.rs + cleanup.rs (split out cleanup utility)
- New util.rs for shared helpers (collect_pinned_func_ids, etc.)

Naming fixes:
- nullable_ref_optimize → optimize_nullable_refs (verb-first)
- drve_dead_return_values → eliminate_dead_return_values (drop acronym)
- optimize_instrs → run_peephole (more specific)
- cleanup_wir → cleanup (module name provides context)
- elide_dead_*_struct_locals → elide_*_struct_locals (drop redundant "dead")

Pipeline restructured with phase comments for clarity.
